+++ bciers/apps/administration-e2e/tests/workflows/industry_user/select-operator-with-beforeAll.spec.ts
@@ -43,7 +43,6 @@ test.describe("Test select operator paths with beforeAll", () => {
   test("Select operator link from dashboard", async ({ page }) => {
     // 🛸 Navigate to dashboard page
     const dashboardPage = new DashboardPOM(page);
-    debugger;
     await dashboardPage.route();
     // 🔍 Assert current URL
     await dashboardPage.urlIsCorrect();
